feat(core): undecorated-classes-with-decorated-fields migration should handle classes with lifecycle hooks
As of v10, we plan in removing the undecorated classes compatibility
code in ngtsc. This means that we will enforce the new mental model
for applying decorators.

We had a migration for this in v9, but it looks like the migration
did not handle the case where a class uses lifecycle hooks. This
is handled in the ngtsc compatibility code, so we should handle
it in migrations too. This has not been outlined in the migration
plan initially, but an appendix has been added for v10.

https://hackmd.io/vuQfavzfRG6KUCtU7oK_EA?both.

Note: This could also mean that we accidentally add `@Directive()`
to undecorated classes which are services using `ngOnDestroy`. There
doesn't seem to be a good way how we can differentiate here. ngtsc
also generates for them a directive definition currently.