feat(core): undecorated-classes-with-di migration should handle inherited constructor for pipes
Pipes could inherit a constructor from undecorated base classes. This is
something that has stopped working with Ivy, but worked previously in
View Engine as long as the necessary TS metadata has been emitted.

We want to handle `@Pipe`'s in the undecorated-classes-with-di migration
so that we can enforce a consistent mental model on where to apply decorators.