feat(compiler-cli): error if declarations inherit from undecorated base classes
As of verison 9, classes should always be decorated if they use Angular
features, or inherit Angular behavior.

This has been done so that there is a single, simpler rule on where
decorators need to be applied in Angular. This behavior will now be
enforced in version 10. ngtsc will report errors:

1. If a declaration inherits a constructor from an undecorated base
   class (Angular needs a decorator so that DI can work)
2. If an undecorated class extends a decorated class. Undecorated
   children are not supported. See v9 migration guide.
3. If a injectable inherits from a directive, component or pipe. This
   is not supported as DI could work unexpectedly (e.g. node injector)
4. If a pipe, directive or injectable inherits from a component. The
   inherit definition feature will already throw for this at runtime, but
   we could also detect it at build time given that we check inheritance too.