refactor(platform-server): switch from dynamic require to import statement
In order to support ESM for the `platform-server` package, we need to
remove two usages of dynamic imports and replace them with their
corresponding/equivalent import statement. This will also allow rollup
to recognize this import but we want to keep these imports external, so
we explicitly specify the `externals` option for the `ng_package` rule.