build: only ship type definitions along with bundles for compiler-cli
We switched the build output for the compiler-cli to use esbuild-generated
bundles.

This means that actual devmode ES5 sources, or prodmode ES2020 non-bundled
sources are not actually needed. Only the types are needed, and this commit
makes sure only the type definitions are shipped. This reduces the code size
of the compiler-cli and also helps with avoiding incorrect module resolution.