build: update to snapshot build of Angular CLI to support ESM package output
The CLI has been updated with various changes to support for ESM package
output of the framework. This commit updates to a snapshot build of the
CLI that contains these necessary changes, allowing us to test/validate
our package changes through integration tests or AIO examples.

This update also contains a fix for the adjust enum babel optimization
that broke with the latest rollup version.