refactor: remove remaining dynamic require usages in package output
Removes the remaining usages of dynamic require statements in the
package output. Since we declare all shipped packages as strict ESM,
we cannot use dynamic require statements anymore. This commit switches
these usages to actual `import` statements.

Note: Tsickle technically is an optional dependency. This change always
imports it regardless of whether there is
`annotateForClosureCompiler` enabled. This is necessary as otherwise
we'd need to use a dynamic import which would make the whole compilation
pipeline async. This is not desirable.

Importing tsickle always is not an issue because we bundle it into the
`ngc` binary anyway. It seems to not contribute a significant size here.
We can revisit removing tsickle from the pipeline here, but for now that
is out of scope/and needs some more discussion because the annotate for
closure compiler API seems somewhat "official" (given it being respected
by `ngtsc` in some places as well)