refactor(compiler-cli): do not use `__filename` global for ESM compatibility
Switches the compiler-cli usage of `__filename` to `import.meta.url`
when ESM bundles are generated. Unfortunately we cannot start using
only `import.meta` yet as we still build and run all code in Angular
in CommonJS module output for devmode tests.

We temporarily replace `__filename` in the compiler-cli using the
rollup replace plugin. This can be removed in the future but unblocks
our ESM support for now.