ci: move zone.js package.json ownership to `dev-infra`.
Renovate frequently updates the `devDependencies` in the `zone.js` package.json
file. This means that a lot of renovate bump version PRs end up requiring review
from `zone-js`. This is a potential bottleneck.

By moving ownership of this file to dev-infra, these updates can be handled
betterr. The downside is that dev-infra owns a file that is part of
the actual `zone.js` public NPM package.