test: revert symbol golden updates after tree-shaking fixes
We've fixed tree-shaking in the `main` branch given the reactive
nodes being considered side-effects due to the spread assignment.

We fixes this in `main` and also applied the fix for our new reactive
node additions- i.e. input signal and queries.

This commit reflects this fix. It's basically a revert of
45dea51bf7d92d396640c1ab84b1b48010e7be4a